+++ .gitignore
@@ -1,5 +1,7 @@
 test
 *.o
+*.d
 build
 bin
+lib
 a.out
