--- armednn.h
@@ -1,16 +1,0 @@
-
-#ifndef _ARMEDNN_H_
-#define _ARMEDNN_H_
-
-#include "common.h"
-#include "operator.h"
-#include "config.h"
-#include "param.h"
-#include "op_regist.h"
-#include "node.h"
-#include "dense.h"
-#include "split.h"
-#include "concat.h"
-#include "fpooling.h"
-
-#endif
