+++ include/armednn.h
@@ -1,0 +1,14 @@
+
+#ifndef _ARMEDNN_H_
+#define _ARMEDNN_H_
+
+#include "armednn/common.h"
+#include "armednn/config.h"
+#include "armednn/param.h"
+#include "armednn/op_regist.h"
+#include "armednn/op_sub.h"
+#include "armednn/operator.h"
+#include "armednn/node.h"
+#include "armednn/special.h"
+
+#endif
