+++ include/armednn/common.h
@@ -1,10 +1,12 @@
-#ifndef _CORE_H
-#define _CORE_H
+#ifndef _COMMOM_H
+#define _COMMOM_H
 
 #include <iostream>
 #include <unordered_map>
 #include <string>
 #include <chrono>
+
+
 #include "Eigen/Dense"
