+++ include/armednn/operator.h
