+++ src/core/op_sub.cpp
@@ -1,6 +1,6 @@
 
 #include <map>
-#include "op_sub.h"
+#include "armednn/op_sub.h"
