--- tests/Makefile
@@ -1,3 +1,0 @@
-
-all:
-	g++ -std=c++11 0_data.cpp -msse4.2 -O2 -I.. -I../third_party/eigen ../node.cpp ../common.cpp ../op_sub.cpp -DEIGEN_RUNTIME_NO_MALLOC
